refactor: RandomNumberGenerator 클래스를 service 패키지에서 util 패키지로 이동
