+++ src/main/java/racingcar/util/RandomNumberGenerator.java
@@ -1,4 +1,4 @@
-package racingcar.service;
+package racingcar.util;
 
 import java.util.Random;
